updates 20170226

+++ 00checksum.sha256.gen.bash
@@ -1,5 +1,3 @@
 #!/usr/bin/env bash
-if [ -f CHECKSUMS.SHA256 ] ; then
-    rm CHECKSUMS.SHA256
-fi
+if [ -f CHECKSUMS.SHA256 ] ; then rm CHECKSUMS.SHA256 ; fi
 sha256sum add.*/* hpHosts*/* rlwpx.free.fr.h*/* *Ads/* > CHECKSUMS.SHA256
